+++ src/components/feed-display/FeedItemDisplay.vue
@@ -11,10 +11,21 @@
 					</template>
 				</NcActionButton>
 			</NcActions>
-			<StarIcon :class="{'starred': item.starred }" tabindex="0" @click="toggleStarred(item)" />
-			<EyeIcon v-if="item.unread" tabindex="0" @click="toggleRead(item)" />
-			<EyeCheckIcon v-if="!item.unread" tabindex="0" @click="toggleRead(item)" />
-			<CloseIcon tabindex="0" @click="compactMode ? $emit('show-details') : clearSelected()" />
+			<StarIcon :class="{'starred': item.starred }"
+				:title="t('news', 'Toggle star article')"
+				tabindex="0"
+				@click="toggleStarred(item)" />
+			<EyeIcon v-if="item.unread"
+				:title="t('news', 'Mark read')"
+				tabindex="0"
+				@click="toggleRead(item)" />
+			<EyeCheckIcon v-if="!item.unread"
+				:title="t('news', 'Mark unread')"
+				tabindex="0"
+				@click="toggleRead(item)" />
+			<CloseIcon :title="t('news', 'Close details')"
+				tabindex="0"
+				@click="compactMode ? $emit('show-details') : clearSelected()" />
 			<button v-if="compactMode"
 				v-shortkey="{s: ['s'], l: ['l'], i: ['i']}"
 				class="hidden"